backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.6

# Task Queue
celery==5.3.6
redis==5.0.1

# STT
faster-whisper>=1.0.0
pyannote.audio==3.1.1

# LLM
google-generativeai==0.3.2
openai==1.10.0

# File handling
python-magic==0.4.27
aiofiles==23.2.1

# Security
cryptography==42.0.1

# Monitoring
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
